Recast exception.
If location() throws a ClassNotFoundException, then the server returned something other than what it was supposed to, so that can be lumped in as a ProtocolException. The user of the ProxyControllerPipeline probably doesn't want to be bothered with the particulars of how the protocol was violated. (I might should log it, though.)