Added "perfect" RNGs.
Added prng-perfect-double-seq, which theoretically uses all bits of precision in all numbers evenly distributed between 0 and 1, so numbers closer to zero have higher resolution. (It's debatable whether this is really better than prng-double-seq.) With the PRNG I'm using, the sequence probably only covers less than 1/64 of the possible range, with the vast majority of the unused parts of the range being near zero.

I don't have any plans to use this code, and probably should never have written it, but it was a fun exercise. More useful would be to try to develop a faster PRNG with similarly small state and good statistical properties, but really this SSLFSR is probably fast enough for my application, which is likely to be IO bound, that it won't ever matter.