Some work on startup; added WorldProxy
Added a minimal WorldProxy class.
Broke out dm-startup! function that opens a database connection based on the parameters given, and starts the writer thread.
Added a startup! function for the server itself.
Fixed rests, which had broken unnoticed in the lazy-seq update. (Unit test much?)
Changed add-self-to-world! to reflect idea that an avatar object is permanent.