Work-in-progress on datastore.
Moved SQL connection opening and CREATE TABLE out of durable-maps and into datastore.
Moved some functions out of server-lib and into a more generally applicable dvtp-lib.
Renamed datastore functions with bangs. (All of them? Is that right?)
Added util/firstarg, for method dispatching on the first argument's _value_, not class, nor a property of it.