Fixed some dm and node-tree design issues.
In durable-maps, die during the attempt to add a write to the write queue that would cause the writer thread to die. (This is still not enough -- the writer thread is still fragile.)
Valify the keyval in datastore/get-row.
Added sphere, vec-abs, room?, node-depth functions in server-lib.
Renamed is-ephem? to just ephem?.
Commented out aborted make-concrete; going to write the real version next and then use it.
Finished initial gen-simple-starsystem.
Changed specs to contain the symbols of :generator functions rather than the functions themselves, which can't be stored in the database.
Added util/exists?.