Added print-dup for DvtpExternalizable.
When checking that a write query gives the key column, allow queries that don't have a row at all.
Added base-64 encoding and decoding, and used that together with the binary externalization format to implement a print-dup for DvtpExternalizable objects.
Using the prettyPrint method, along with the #<> unreadable convention, for a print-method for DvtpExternalizable objects.
Got gen-children generating some node-like things, and then immediately broke it again as a step toward making them consistent with the root node I added manually.
Added full class names (and constuctor dots) to prettyPrint() methods as part of an aborted attempt to turn that into a print-dup format.
Normalize ints when looking up nodes. (Still need to do this when adding them.)