Got initial conversation working.
Fixed add-self-to-world and start-rendering.
Replaced invalid tables with a function that throws an informative exception instead of the string "fail".
Added spaces inside some double open parens -- new coding style standard.
Corrected mortifying misspelling.
Renamed ->> macro to -->> to avoid collision with new core macro.
Added let-when macro.
Moved call to start-rendering outside the transaction in which the avatar is added to the world.