Some work on a static-scene server
Added a comment explaining the meaning of test_spec.clj.
Added message-to-bytes and messages-to-bytes.
Routed multiple servers through a common server command in core.clj.
Added id-to-object map.
Added handling of add-object messages
Bumped all the IDs in the hardcoded scene graph so ID 0 doesn't exist (so
nodes can be added without existing in the scene graph anywhere).