WIP on movement
Added several protocol classes to support movement; none of them written in
stone yet.
Changed add-object class to use :moves instead of :pos.
Converted all :posses to :moveses in static scenes (in client and static
server).
Added utility functions to ease transition to moveseqs.

Still need to add a notion of time in the client, and to use :moves instead
of :pos when rendering.